Tools: Removed typedoc from gulp tasks.
