+++ tools/gulptasks/tsdoc.js
@@ -3,7 +3,6 @@
  */
 
 const gulp = require('gulp');
-const path = require('path');
 
 /* *
  *
@@ -11,20 +10,12 @@ const path = require('path');
  *
  * */
 
-const INTERNAL_TARGET = path.join('build', 'api-internals');
-/*
-const INTERNAL_THEME = path.join(
-    'node_modules', '@highcharts', 'highcharts-documentation-generators',
-    'typedoc', 'theme-internals'
+const TREE_TARGET = 'tree-next.json';
+
+const TEMPLATE_SOURCE = (
+    'node_modules/@highcharts/highcharts-documentation-generators/templates'
 );
- */
-const NEXT_TARGET = path.join('build', 'api-next');
-/*
-const NEXT_THEME = path.join(
-    'node_modules', '@highcharts', 'highcharts-documentation-generators',
-    'typedoc', 'theme-next'
-);
- */
+
 /* *
  *
  *  Tasks
@@ -32,53 +23,64 @@ const NEXT_THEME = path.join(
  * */
 
 /**
- * TSDoc task
- * @return {Promise<undefined>}
+ * TSDoc-next task
+ * @return {Promise<void>}
  *         Promise to keep
  */
 function task() {
 
-    const argv = require('yargs').argv;
-    const fsLib = require('./lib/fs');
-    const processLib = require('./lib/process');
-
-    const folders = fsLib
-        .getDirectoryPaths('ts')
-        .map(folder => folder.substring(3))
-        .filter(folder => !folder.startsWith('masters'));
-    const target = argv.next ? NEXT_TARGET : INTERNAL_TARGET;
-    // const theme = argv.next ? NEXT_THEME : INTERNAL_THEME;
-
-    const command = (
-        'cd ts && npx typedoc' +
-        ` {${folders}}/{*,**/*,**/**/*,**/**/**/*} ` +
-        ` --json "${path.join('..', target, 'tree.json')}"` +
-        ` --out "${path.join('..', target)}"`
-        // + ` --theme "${path.join('..', theme)}"`
-    );
-
-    return processLib
-        .exec(command)
-        .then(() => processLib.openAppFor(
-            path.join((argv.next ? NEXT_TARGET : INTERNAL_TARGET), 'index.html')
-        ));
-
-    /*
+    const args = require('yargs').argv;
+    const fs = require('fs');
     const generators = require(
         '@highcharts/highcharts-documentation-generators'
     );
-    const log = require('./lib/log');
+    const logLib = require('./lib/log');
 
-    return new Promise(resolve => {
+    if (args.old) {
+        const generator = generators.TypeScript4;
 
-        log.message('Generating', TARGET_DIRECTORY + '...');
+        return Promise
+            .resolve()
+            .then(() => generator.ProjectDoc.load('ts'))
+            .then(project => fs.promises.writeFile(
+                TREE_TARGET,
+                generator.JSON.stringify(project.toJSON())
+            ))
+            .then(logLib.success)
+            .catch(logLib.failure);
+    }
 
-        generators.TypeDoc
-            .task(SOURCE_CONFIG, TARGET_DIRECTORY, TARGET_JSON)
-            .then(() => log.success('Created', TARGET_DIRECTORY))
-            .then(resolve);
-    });
-    */
+    const generator = generators.Generator;
+    const parser = generators.Parser;
+    const source = (
+        fs.existsSync('code/es-modules/') ?
+            'code/es-modules/' :
+            'ts/'
+    );
+
+    if (source === 'ts/') {
+        logLib.warn('ES modules not found, using ./ts/ folder.');
+    }
+
+    return parser.Project
+        .load('ts', args)
+        .then(project => {
+            fs.promises.writeFile(
+                TREE_TARGET,
+                parser.JSON.stringify(project.toJSON())
+            );
+        })
+        .then(() => generator.Template.loadFolder(TEMPLATE_SOURCE))
+        .then(() => generator.Project.load(TREE_TARGET))
+        .then(project => {
+            generator.Template.types.classes.write(
+                'build/classes.html',
+                project.ast
+            );
+        })
+        .then(logLib.success)
+        .catch(error => logLib.failure(error, error.stack));
+
 }
 
-gulp.task('tsdoc', gulp.series('clean-api', task));
+gulp.task('tsdoc', task);
